Fixed failing test on i386 due to floating point inaccuracies.
Calculating the numer of passes give the incorrect result on i386 when
using double.  Switching to long double avoid the problem.

Fixes #5822.

+++ kernel/fstdata.h
@@ -55,7 +55,7 @@ class FstData
 	std::string valueOf(fstHandle signal);
 	fstHandle getHandle(std::string name);
 	dict<int,fstHandle> getMemoryHandles(std::string name);
-	double getTimescale() { return timescale; }
+	long double getTimescale() { return timescale; }
 	const char *getTimescaleString() { return timescale_str.c_str(); }
 private:
 	void extractVarNames();
@@ -69,7 +69,7 @@ private:
 	uint64_t last_time;
 	std::map<fstHandle, std::string> past_data;
 	uint64_t past_time;
-	double timescale;
+	long double timescale;
 	std::string timescale_str;
 	uint64_t start_time;
 	uint64_t end_time;
